[TASK] Use MAC activity for PLK LED
Instead of using the PHY generated activity blinking, the activity
generated by openMAC is used. The PHY link is observed from iLinkPlkPhy
port - for this the LED mode has to be changed that the link port is
asserted when a link is detected (no blinking in case of activity). This
is changed in omethlib_phycfg.c at EDRV startup.

Change-Id: Ic7adc62c656c5306683506e3f0cf99b98cd635e2