[FIX] Check for update image also after reconfiguration
Reconfiguring the FPGA from update image mode back to update image
again does not work with Cyclone 4 devices. Instead the FPGA
reconfigures to factory image mode. This is the reason why the firmware
update fails if the FPGA is in update image mode.
In order to workaround this limitation the factory image checks for a
valid update image after POR and also after reconfiguration.

Change-Id: Ib77adf9f4e89c52a7f3a0559cbe515d92fd15406