[FIX] Fix stack cleanup and shutdown command handling
The firmware update application does not issue the kernel stack
initialization, but it commands to shut down the stack afterwards. This
could lead to unfreed memory buffers due to the production test module
(which uses edrv module of stack).
This fix introduces handling of cleanup and shutdown commands if the
stack is not initialized.

Change-Id: Iea5f4cc7f8443140e963b1881b96217c6a849e63