[FIX] Connect atomic ipcore directly to shared memory
The atomic ipcore was instantiated in the pcie subsystem and not
directly connected with the memory controller. Thus, the locking signal
vital for atomic modification operation was not mapped to the memory
controller.
With this fix the atomic modify ipcore is removed from the subsystem and
instantiated in the toplevel qsys directly connected to the memory
controller for proper operation.

Change-Id: I978a95178624d0e113595e885914737ef5f92d26